fz.ias_smoke_alarm_1_develco and fz.WSZ01_on_off_action
