+++ src/devices/develco.ts
@@ -179,6 +179,20 @@ const develco = {
                 return result;
             },
         } satisfies Fz.Converter<"haElectricalMeasurement", undefined, ["attributeReport", "readResponse"]>,
+        ias_smoke_alarm_1_develco: {
+            cluster: "ssIasZone",
+            type: "commandStatusChangeNotification",
+            convert: (model, msg, publish, options, meta) => {
+                const zoneStatus = msg.data.zonestatus;
+                return {
+                    smoke: (zoneStatus & 1) > 0,
+                    battery_low: (zoneStatus & (1 << 3)) > 0,
+                    supervision_reports: (zoneStatus & (1 << 4)) > 0,
+                    restore_reports: (zoneStatus & (1 << 5)) > 0,
+                    test: (zoneStatus & (1 << 8)) > 0,
+                };
+            },
+        } satisfies Fz.Converter<"ssIasZone", undefined, "commandStatusChangeNotification">,
     },
     tz: {
         pulse_configuration: {
@@ -404,7 +418,7 @@ export const definitions: DefinitionWithExtend[] = [
             {vendor: "Frient", model: "94430", description: "Smart Intelligent Smoke Alarm"},
             {vendor: "Cavius", model: "2103", description: "RF SMOKE ALARM, 5 YEAR 65MM"},
         ],
-        fromZigbee: [fz.ias_smoke_alarm_1_develco, fz.ias_enroll, fz.ias_wd, develco.fz.fault_status],
+        fromZigbee: [develco.fz.ias_smoke_alarm_1_develco, fz.ias_enroll, fz.ias_wd, develco.fz.fault_status],
         toZigbee: [tz.warning, tz.ias_max_duration, tz.warning_simple],
         ota: true,
         extend: [
@@ -481,7 +495,7 @@ export const definitions: DefinitionWithExtend[] = [
         vendor: "Develco",
         description: "Fire detector with siren",
         whiteLabel: [{vendor: "Frient", model: "94431", description: "Smart Intelligent Heat Alarm"}],
-        fromZigbee: [fz.ias_smoke_alarm_1_develco, fz.ias_enroll, fz.ias_wd, develco.fz.fault_status],
+        fromZigbee: [develco.fz.ias_smoke_alarm_1_develco, fz.ias_enroll, fz.ias_wd, develco.fz.fault_status],
         toZigbee: [tz.warning, tz.ias_max_duration, tz.warning_simple],
         ota: true,
         extend: [
